Significant minor changes to improve "code smells" and the start of additional test cases around notifications.
Refactoring of how energy types are handled which simplifies their usage and units conversion.

+++ src/metering/common/tiers.cpp
@@ -45,7 +45,7 @@ Tiers::SupportedTiers Tiers::FromString(const std::string& pricing_tier_string)
 		break;
 
 	case V2_LENGTH:		// Format: '0x12'
-		tier_as_int = std::stoul(pricing_tier_string, nullptr, 16);
+		tier_as_int = static_cast<uint8_t>(std::stoul(pricing_tier_string, nullptr, 16));
 		break;
 
 	default:
